kernel-2.6.18-238.9.1.el5
CentOS kernel, generated by rpmbuild -bp --target=noarch

+++ arch/x86_64/kernel/smpboot.c
@@ -942,7 +942,7 @@ do_rest:
 	 */
 	CMOS_WRITE(0, 0xf);
 
-	*((volatile long *) phys_to_virt(0x467)) = 0;
+	*((volatile u32 *) phys_to_virt(0x467)) = 0;
 
 	return 0;
 }
